commit Slowly add and subtract gears

+++ Src/Drivers/dac8831.c
@@ -173,16 +173,14 @@ void Wave_select(uint8_t n, uint16_t *_pBuf)
 	}
 }
 
-void Dac8831_Set_Amp(uint16_t amp, uint16_t *_pBuf)
+void Dac8831_Set_Amp(float amp, uint16_t *_pBuf)
 {
 	int buf_data;
 	for(int i = 0; i < 100; ++i)
 	{
 		buf_data = _pBuf[i];
 		buf_data = (buf_data - 32767)*amp / 255 + 32767;
-		_pBuf[i] = (uint16_t)buf_data;
-		
-		
+		_pBuf[i] = (uint16_t)buf_data;	
 	}
 	 
 }
